rouille is not outdated
The rouille server-side framework is not outdated. It is actively
maintained (the last release was less than 3 weeks ago as of the time
of this commit), and it is used in a wide variety of public and
private projects (some of which I maintain!).

rouille does not use async Rust, but that is in fact the point of the
framework, as it is intended for use-cases where the costs of async
are not balanced by its upsides.

This fixes #134.