Add markup and tinytemplate, remove yew-template
`markup` and `tinytemplate` are two incredibly easy to use, and well
performing template engines. `yew-template` is not a template engine,
but an add-on to `yew`.

I marked both new engines as `outdated = true`, because they did not
receive any updates recently, but I guess I would rather call them
"mature" or "feature complete".